+++ frontend/libs/identity/identity-ui/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/roles.page';
+export * from './lib/users.page';